Got the textPane implemented.
This is basically the full GUI. Now all that needs to be done is to add
the buttons and other feature but first i need to know how the game
works to do that (kind of). Put border all around the main components to
make it look nice. Things to do next is to get the right click pop up
menus working. Probably not a bad idea to go through and move some parts
of the main into helper methods. Ill do that later if it will get more
marks for code style, otherwise DGAF